Mitigated json-server nested resource foreign key write bug for POST help request
The bug is that the id parameter taken from the url, somewhere deep down inside was being treated as a string without parsing it to number. Hence the saved entity has the foreign key as a string and it ends up failing to match any internal query upon doing GET.

Tried changing the value within params object, but it seems to have no effect, apparently internally, the value is taken from the url - so I had to change that. Also I had to manually add the foreign key, since it waz being done in a buggy manner via url, which I had to override.