Added case notes embeding upon querying specific resident
In case you're wondering, whenever PUT is being routed to a url with "_embed" parameter, that parameter gets ignored, and the PUT works just fine. And it does not delete the collection due to this being structured in a relational DB fashion rather than MongoDB way. If it was mongodb way, where case notes array were literally a property of residents object, then yes, it could be deleted maybe... since as far as I check, if you don't specify some field in PUT, the update version is missing it.