Added help request calls embeding upon querying help requests
It's a simple route map to an _embed, which doesn 't work due to json-server bug, however this route gets intercepted via middleware function that was editted last commit.

Additional information: Whenever POST gets routed to this url, the _embed gets ignored and does not affect the endpoint. The POST with _embed works as if _embed was not even there. And the middleware mentioned before is an outbound middleware that only affects a request if it was a GET.

Also remove an interesting looking console.log() :D Pardon me, but it's literally midnight and stuff wasn't working properly for a really long time.