Made assign page accept additional "resident_id" parameter
Had to do so that the request to the backend could be made as the agreed url to access help_request is /resident/:resident_id/help_requests/:help_request_id
(even though the resident id is not necessary to do so (we're using the rest naming convention))

There was an option to not include the resident_id in the url string by simply having it in the query parameters object. This way it could get accessed without it needing to be in url. However, I had to make it so resident Id is displayed in the url string so that the Reassign Single call page would work after refresh (as after refresh the only thing that stays arround is url, so the resident_id would get lost, and hence the help_request retrieval would fail due to missing parameter). Also made sure that the resident_id is in the query string so that I could avoid creating a mess of a folder structure just so it could be represented in the REST way.