Magic that makes the router.query retrieve params from url after Refresh!
Got no clue how or why it works, but it does - the magic line!

The problem was that the page would work fine if you navigated to it from the callbacks list page, however it would fail to retrieve the route parameters upon navigating to the url directly with the parameters explicitly in the url. Due to the parameters being undefined, an API call would fail, also the page would be displayed as 404 not found.