Moved rescheduled_at & requested_date fields to Help Request (mockapi)
Moved them so that this data wouldn't be random every time I refresh the callbacks page - it was stupid of me to implement it that way to begin with & it interferes now with a thing I want to test.

When I implemented it as random every requested, I was hoping it will get resolved soon... welp no... this thing is more of a can kicked down the road. Will get to it eventually, once the backend progresses to the point where they finish data migration & start writing endpoints.