Added Reassign Single Call page Reload test
The page behaving correctly after reload does not come out of the box, hence a test to cover the coded added to make it work has to be tested.

Under refresh conditions the page loads, but the query parameters from url do not get passed into an API call (on time at least), making the page fail to retrieve the Help Request that would allow to set the correct assignee dropdown value (and update it in DB later on).