Added case notes gateway test checking for the invocation of mapper
A somewhat hacky way of setting up mapper mock (modifying the global instance of the mapper class method), but it gets the job done, and it's not much worse than what moxios is already doing (modifying the responses of global axios instance). To mitigate this global noise nonsense, I made sure the tests run sync, even though they are calling async functions, and are marked as async themselves.