Made "checked" attribute appear conditionally based on what page is using CallbackForm component
If page is addSupport, then helpRequestExists = false, and if it's manageRequest, the it is true (and needs fetched values to be populated).

Have to use 2 separate Radio button groups, because I want to use "checked" rather than "defaultChecked" - reason for it is that "checked" allows for locking down the selected fetched radio button due to checked requiring explicit control over state. If I were to use defaultChecked, thenthe radio buttons could be misleadingly reasigned to some other radio button value on UI, even though it won't get saved to DB.