Moved out the setting of initial state from useState to useEffect on CallbackForm
Reason is that ... the way how the state will be set will depend on which page is going to use this form component. If it's the addSupport page, then helpRequest is going to be undefined, and hence we want to resort to some default values. If it's the manageRequest page, then we want to use values from the API call that the parent page component (of the CallbackForm) did.

This is where the async and multi-level state madness starts, but in short, we need to use useEffect so that we could redraw CallbackForm once the data gets fetched.

Have to specify helpRequest in dependency array, because it's being retrieved via props before the the async messy state setting is done... so it initially gets passed only an empty object. However once the async madness finishes up, the n the parent's helpRequest state variable gets update, which triggers props update, which with the help of [helpRequest] within useEffect triggers component's internal state update, which is finally needed to draw the rest of the UI properly. Jesus christ...