Added onChange handler for address lines fields
So that they could be saved by entering data manually rather than populating it via the dropdown.

It so happens that Addresses API can't find all of the postcodes & hence not all addresses can be entered auto-populated. In this case the User tries to enter this data manually, but the code to capture this manually entered was not written, as it was expected that Addresses database is exhaustive.