Scroll up when 'CallbackForm' validation fails.
Prior to this commit the validation failure was uneventful. It seemed as if the button didn't even work. Apparently, at the top of the page a generic validation error message already exists. So I added JavaScript to scroll the user to it. Of course, ideal fix would be to have each component complain for themselves about them not being filled in; however, that's also a refactoring tech-debt ticket, so it's left for later.