This if / Else If / Else was confused. It was simultaneously checking the selectedAssignment (checkboxes selected) and whether that selection had any data.
I removed the data check as it wasn't required, as long as the array is empty and not null the code will function and do nothing.

With these checks, it would fall into the final Else statement if someone selected 'Unassigned' and there were no unassigned users, it would then proceed to assign already-assigned users again, even though it hadn't been requested.