Fix asynchronous function calls within the UseEffect.
This fixes all of those 'destroy is not a function' errors.

If the function executed by useEffect return anything, the useEffect assumes its a destroy (clean up) function.

The issue is that when you make an async function call from the useEffect dirrectly (no anon function wrapper), then that function return a Promise.

The useEffect tries to execute that promise as 'destroy' function & runs into runtime exception of: 'destroy is not a function'.

Don't know why, but the previous version of react or the framework didn't have an issue with this. Now it does, hence the change.